+++ app/schemas/nutrition.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel
 class NutritionSchema(BaseModel):
     user_telegram_id: int
     calories: int
-    water: int
+    water: float
 
     class Config:
         orm_mode = True
